+++ python/tests/functional/eyes_images/test_images.py
@@ -1,3 +1,4 @@
+import json
 import os
 
 import pytest
@@ -91,6 +92,22 @@ def test_image_extract_text_region(ocr_image, eyes):
     text = eyes.extract_text(OCRRegion(ocr_image, Region(55, 11, 214, 18)))
 
     assert text == ["s the navigation bar"]
+
+
+def test_images_send_dom(png_target, eyes, app_output):
+    dom = {
+        "childNodes": [],
+        "css": "",
+        "images": {},
+        "rect": {"height": 600, "left": 0, "top": 0, "width": 800},
+        "scriptVersion": "11.2.1",
+        "style": {},
+        "tagName": "HTML",
+        "version": "1.3.0",
+    }
+    eyes.check(png_target.dom(json.dumps(dom)))
+
+    assert app_output()[0]["image"]["hasDom"]
 
 
 def img_path(kind):